Make MotorPair methods a bit more robust against being unpaired
Flipper by and large signals a pair that has become unpaired by
returning False from method calls on that pair.  This isn't a
good strategy (raising an exception would be a lot better), but
for now let's follow the lead of the original.

The exception to this rule is that MotorPair.primary() and
 .secondary() still return the original Motor objects, *even if
the corresponding motor has been disconnected.*  This seems
unnecessarily dangerous, and Strawberry raises an exception
instead.