Call the Motor or MotorPair callbacks.
The callbacks are called on receipt of an Output Feedback
message, i.e. from the background thread.  Calling any
methods that send messages is likely to go badly wrong
at this point; the queue is not robust, and deadlock would
be easy to achieve accidentally.  This needs some design
thought.